fix(editor): Move project types
