+++ packages/editor-ui/src/components/CredentialEdit/CredentialSharing.ee.vue
@@ -85,18 +85,18 @@ import { useUsageStore } from '@/stores/usage.store';
 import { EnterpriseEditionFeature } from '@/constants';
 import ProjectSharing from '@/features/projects/components/ProjectSharing.vue';
 import { useProjectsStore } from '@/features/projects/projects.store';
-import type {
+import {
 	ProjectListItem,
 	ProjectSharingData,
 	Project,
-} from '@/features/projects/projects.types';
+	ProjectTypes,
+} from '@/types/projects.types';
 import type { ICredentialDataDecryptedObject } from 'n8n-workflow';
 import type { PermissionsMap } from '@/permissions';
 import type { CredentialScope } from '@n8n/permissions';
 import type { EventBus } from 'n8n-design-system/utils';
 import { useRolesStore } from '@/stores/roles.store';
 import type { RoleMap } from '@/types/roles.types';
-import { ProjectTypes } from '@/utils/projects.utils';
 
 export default defineComponent({
 	name: 'CredentialSharing',
