+++ packages/editor-ui/src/components/DeleteUserModal.vue
@@ -84,7 +84,7 @@ import { mapStores } from 'pinia';
 import { useUsersStore } from '@/stores/users.store';
 import { createEventBus } from 'n8n-design-system/utils';
 import { useProjectsStore } from '@/features/projects/projects.store';
-import type { ProjectListItem, ProjectSharingData } from '@/features/projects/projects.types';
+import type { ProjectListItem, ProjectSharingData } from '@/types/projects.types';
 import ProjectSharing from '@/features/projects/components/ProjectSharing.vue';
 
 export default defineComponent({
