+++ packages/editor-ui/src/components/__tests__/CredentialCard.test.ts
@@ -6,7 +6,7 @@ import { useUIStore } from '@/stores/ui.store';
 import { useUsersStore } from '@/stores/users.store';
 import { useCredentialsStore } from '@/stores/credentials.store';
 import type { ICredentialsResponse } from '@/Interface';
-import type { ProjectSharingData } from '@/features/projects/projects.types';
+import type { ProjectSharingData } from '@/types/projects.types';
 
 const renderComponent = createComponentRenderer(CredentialCard);
