+++ packages/editor-ui/src/stores/credentials.store.ts
@@ -35,7 +35,7 @@ import { useRootStore } from './n8nRoot.store';
 import { useNodeTypesStore } from './nodeTypes.store';
 import { useSettingsStore } from './settings.store';
 import { isEmpty } from '@/utils/typesUtils';
-import type { ProjectSharingData } from '@/features/projects/projects.types';
+import type { ProjectSharingData } from '@/types/projects.types';
 import { splitName } from '@/utils/projects.utils';
 
 const DEFAULT_CREDENTIAL_NAME = 'Unnamed credential';
